cleaned up carriers list. fixed active/renewal in plans. added common header object with phone/sms/email/location

+++ HBXMobile/benefitGroupCardView.h
@@ -42,6 +42,6 @@
 
 @property (nonatomic, weak)id<planCardViewDelegate> delegate;
 
--(void)layoutView:(int)cp totalPages:(int)tp;
+-(void)layoutView:(int)cp totalPages:(NSInteger)tp;
 
 @end
